+++ src/pipx/main.py
@@ -240,7 +240,9 @@ def run_pipx_command(args: argparse.Namespace) -> ExitCode:  # noqa: C901
             force=args.force,
         )
     elif args.command == "list":
-        return commands.list_packages(venv_container, args.include_injected, args.json)
+        return commands.list_packages(
+            venv_container, args.include_injected, args.json, args.short
+        )
     elif args.command == "uninstall":
         return commands.uninstall(venv_dir, constants.LOCAL_BIN_DIR, verbose)
     elif args.command == "uninstall-all":
@@ -499,9 +501,11 @@ def _add_list(subparsers: argparse._SubParsersAction) -> None:
         action="store_true",
         help="Show packages injected into the main app's environment",
     )
-    p.add_argument(
+    g = p.add_mutually_exclusive_group()
+    g.add_argument(
         "--json", action="store_true", help="Output rich data in json format."
     )
+    g.add_argument("--short", action="store_true", help="List packages only.")
     p.add_argument("--verbose", action="store_true")