+++ docs/changelog.md
@@ -7,6 +7,7 @@ dev
 - [dev] Change test names
 - Improve the behaviour of `shlex.split` on Windows, so paths on Windows can be handled peoperly when they are passed in `--pip-args`. (#794)
 - Add `pipx environment` command. (#793)
+- Add `list --short` option to list only package names. (#804)
 
   1.0.0